Mejoras

+++ keyboard.c
@@ -31,7 +31,7 @@ void keyboardInit(){
 }
 
 short waitForKeyPress(){
-    int key = -1;
+    short key = -1;
     
     while(key < 0){
         key = get_pressed_key();
